requirements.txt
discord.py
google-cloud-speech
google-cloud-texttospeech
google-cloud-aiplatform
pyaudio
discord.py
numpy
precise-runner